requirements.txt
vllm==0.12.0
datasets>=2.14.0
pandas>=2.0.0
pyarrow>=14.0.0
pyyaml>=6.0
aiohttp>=3.8.0
wandb>=0.15.0
verl[vllm]>=0.7.0
